make codegen better and switch to single-precision

+++ Firmware/src/tests/tracking-test/main.cpp
@@ -17,7 +17,7 @@ void setup() {
 
 void loop() {
     if (Serial.available()) {
-        double gps[3];
+        float gps[3];
         gps[0] = Serial.parseFloat();
         Serial.read();
         gps[1] = Serial.parseFloat();
@@ -26,16 +26,16 @@ void loop() {
         Serial.clear();
         Serial.println("Starting update");
         update = micros();
-        tracker.GPSUpdate((double) (micros() - startTime)/1000000.0, gps);
+        tracker.GPSUpdate((float) (micros() - startTime)/1000000.0, gps);
         update = micros() - update;
         Serial.println(update);
         Serial.println("End update");
     }
 
     if (micros() - lastMessage > 1000 * 1000) {
-        double state[9];
+        float state[9];
         update = micros();
-        tracker.extrapolate((double) (micros() - startTime)/1000000.0, state);
+        tracker.extrapolate((float) (micros() - startTime)/1000000.0, state);
         update = micros() - update;
         Serial.println(update);
         Serial.print(state[0]);
